migrate coro and event loop to seperate files
