+++ include/modules/crypto.h
@@ -1,7 +1,7 @@
 #ifndef CRYPTO_H
 #define CRYPTO_H
 
-#include "ant.h"
+#include "types.h"
 
 void init_crypto_module();
 jsval_t crypto_library(struct js *js);
