+++ include/modules/events.h
@@ -1,7 +1,7 @@
 #ifndef EVENTS_H
 #define EVENTS_H
 
-#include "ant.h"
+#include "types.h"
 
 void init_events_module(void);
 jsval_t events_library(struct js *js);
