+++ include/modules/json.h
@@ -2,7 +2,7 @@
 #define JSON_H
 #define YYJSON_SKIP_VALUE ((yyjson_mut_val *)-1)
 
-#include "ant.h"
+#include "types.h"
 
 void init_json_module(void);
