+++ include/modules/os.h
@@ -1,7 +1,7 @@
 #ifndef ANT_OS_MODULE_H
 #define ANT_OS_MODULE_H
 
-#include "ant.h"
+#include "types.h"
 
 jsval_t os_library(struct js *js);
