+++ include/modules/path.h
@@ -1,7 +1,7 @@
 #ifndef ANT_PATH_MODULE_H
 #define ANT_PATH_MODULE_H
 
-#include "ant.h"
+#include "types.h"
 
 jsval_t path_library(struct js *js);
