+++ include/modules/timer.h
@@ -5,7 +5,6 @@
 #include "types.h"
 
 void init_timer_module(void);
-void process_timers(struct js *js);
 void process_microtasks(struct js *js);
 void process_immediates(struct js *js);
 void queue_microtask(struct js *js, jsval_t callback);
@@ -15,6 +14,5 @@ void timer_gc_update_roots(GC_OP_VAL_ARGS);
 int has_pending_timers(void);
 int has_pending_microtasks(void);
 int has_pending_immediates(void);
-int64_t get_next_timer_timeout(void);
 
 #endif
