+++ include/roots.h
@@ -1,7 +1,7 @@
 #ifndef ROOTS_H
 #define ROOTS_H
 
-#include "ant.h"
+#include "types.h"
 
 jshdl_t js_root(struct js *js, jsval_t val);
 jsval_t js_deref(struct js *js, jshdl_t h);
