+++ include/snapshot.h
@@ -1,10 +1,10 @@
 #ifndef ANT_SNAPSHOT_LOADER_H
 #define ANT_SNAPSHOT_LOADER_H
 
-#include <stdint.h>
-#include "ant.h"
+#include <stddef.h>
+#include "types.h"
 
-jsval_t ant_load_snapshot(struct js *js);
+jsval_t ant_load_snapshot(ant_t *js);
 const uint8_t *ant_get_snapshot_source(size_t *len);
 
 #endif
