+++ src/modules/timer.c
@@ -4,12 +4,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-
-#ifdef _WIN32
-#include <windows.h>
-#else
-#include <sys/time.h>
-#endif
+#include <uv.h>
 
 #include "arena.h"
 #include "errors.h"
@@ -18,13 +13,13 @@
 #include "modules/timer.h"
 
 typedef struct timer_entry {
+  uv_timer_t handle;
   jsval_t callback;
-  uint64_t target_time_ms;
-  uint64_t interval_ms;
   int timer_id;
   int active;
   int is_interval;
   struct timer_entry *next;
+  struct timer_entry *prev;
 } timer_entry_t;
 
 typedef struct microtask_entry {
@@ -49,19 +44,38 @@ static struct {
   immediate_entry_t *immediates_tail;
   int next_timer_id;
   int next_immediate_id;
-} timer_state = {NULL, NULL, NULL, NULL, NULL, NULL, 1, 1};
+  int active_timer_count;
+} timer_state = {NULL, NULL, NULL, NULL, NULL, NULL, 1, 1, 0};
 
-static uint64_t get_current_time_ms(void) {
-#ifdef _WIN32
-  LARGE_INTEGER freq, count;
-  QueryPerformanceFrequency(&freq);
-  QueryPerformanceCounter(&count);
-  return (uint64_t)(count.QuadPart * 1000 / freq.QuadPart);
-#else
-  struct timeval tv;
-  gettimeofday(&tv, NULL);
-  return (uint64_t)tv.tv_sec * 1000 + (uint64_t)tv.tv_usec / 1000;
-#endif
+static void timer_callback(uv_timer_t *handle) {
+  timer_entry_t *entry = (timer_entry_t *)handle->data;
+  if (!entry || !entry->active) return;
+  
+  struct js *js = timer_state.js;
+  jsval_t args[0];
+  js_call(js, entry->callback, args, 0);
+  process_microtasks(js);
+  
+  if (!entry->is_interval) {
+    entry->active = 0;
+    uv_timer_stop(&entry->handle);
+    timer_state.active_timer_count--;
+  }
+}
+
+static void add_timer_entry(timer_entry_t *entry) {
+  entry->next = timer_state.timers;
+  entry->prev = NULL;
+  if (timer_state.timers) {
+    timer_state.timers->prev = entry;
+  }
+  timer_state.timers = entry;
+}
+
+static void remove_timer_entry(timer_entry_t *entry) {
+  if (entry->prev) entry->prev->next = entry->next;
+  else timer_state.timers = entry->next;
+  if (entry->next) entry->next->prev = entry->prev;
 }
 
 // setTimeout(callback, delay)
@@ -72,24 +86,21 @@ static jsval_t js_set_timeout(struct js *js, jsval_t *args, int nargs) {
   
   jsval_t callback = args[0];
   double delay_ms = js_getnum(args[1]);
-  
-  if (delay_ms < 0) {
-    return js_mkerr(js, "setTimeout delay must be non-negative");
-  }
+  if (delay_ms < 0) delay_ms = 0;
   
   timer_entry_t *entry = ant_calloc(sizeof(timer_entry_t));
-  if (entry == NULL) {
-    return js_mkerr(js, "failed to allocate timer");
-  }
+  if (entry == NULL) return js_mkerr(js, "failed to allocate timer");
   
+  uv_timer_init(uv_default_loop(), &entry->handle);
+  entry->handle.data = entry;
   entry->callback = callback;
-  entry->target_time_ms = get_current_time_ms() + (uint64_t)delay_ms;
-  entry->interval_ms = 0;
   entry->timer_id = timer_state.next_timer_id++;
   entry->active = 1;
   entry->is_interval = 0;
-  entry->next = timer_state.timers;
-  timer_state.timers = entry;
+  
+  add_timer_entry(entry);
+  timer_state.active_timer_count++;
+  uv_timer_start(&entry->handle, timer_callback, (uint64_t)delay_ms, 0);
   
   return js_mknum((double)entry->timer_id);
 }
@@ -102,40 +113,35 @@ static jsval_t js_set_interval(struct js *js, jsval_t *args, int nargs) {
   
   jsval_t callback = args[0];
   double delay_ms = js_getnum(args[1]);
-  
-  if (delay_ms < 0) {
-    return js_mkerr(js, "setInterval delay must be non-negative");
-  }
+  if (delay_ms < 0) delay_ms = 0;
   
   timer_entry_t *entry = ant_calloc(sizeof(timer_entry_t));
-  if (entry == NULL) {
-    return js_mkerr(js, "failed to allocate timer");
-  }
+  if (entry == NULL) return js_mkerr(js, "failed to allocate timer");
   
+  uv_timer_init(uv_default_loop(), &entry->handle);
+  entry->handle.data = entry;
   entry->callback = callback;
-  entry->target_time_ms = get_current_time_ms() + (uint64_t)delay_ms;
-  entry->interval_ms = (uint64_t)delay_ms;
   entry->timer_id = timer_state.next_timer_id++;
   entry->active = 1;
   entry->is_interval = 1;
-  entry->next = timer_state.timers;
-  timer_state.timers = entry;
+  
+  add_timer_entry(entry);
+  timer_state.active_timer_count++;
+  uv_timer_start(&entry->handle, timer_callback, (uint64_t)delay_ms, (uint64_t)delay_ms);
   
   return js_mknum((double)entry->timer_id);
 }
 
 // clearTimeout(timerId)
 static jsval_t js_clear_timeout(struct js *js, jsval_t *args, int nargs) {
-  if (nargs < 1) {
-    return js_mkundef();
-  }
-  
+  if (nargs < 1) return js_mkundef();
   int timer_id = (int)js_getnum(args[0]);
   
   for (timer_entry_t *entry = timer_state.timers; entry != NULL; entry = entry->next) {
-    if (entry->timer_id == timer_id) {
+    if (entry->timer_id == timer_id && entry->active) {
       entry->active = 0;
-      break;
+      uv_timer_stop(&entry->handle);
+      timer_state.active_timer_count--; break;
     }
   }
   
@@ -279,76 +285,12 @@ int has_pending_immediates(void) {
   return 0;
 }
 
-static void remove_timer(timer_entry_t *target) {
-  timer_entry_t **ptr = &timer_state.timers;
-  
-scan:
-  if (!*ptr) return;
-  if (*ptr == target) { 
-    *ptr = target->next; 
-    free(target); return; 
-  }
-  ptr = &(*ptr)->next;
-  goto scan;
-}
-
-void process_timers(struct js *js) {
-  uint64_t current_time = get_current_time_ms();
-  timer_entry_t **ptr = &timer_state.timers;
-  timer_entry_t *entry;
-
-scan:
-  if (!*ptr) return;
-  entry = *ptr;
-  
-  if (!entry->active) {
-    *ptr = entry->next;
-    free(entry);
-    goto scan;
-  }
-  
-  if (current_time < entry->target_time_ms) {
-    ptr = &entry->next;
-    goto scan;
-  }
-  
-  jsval_t args[0];
-  js_call(js, entry->callback, args, 0);
-  process_microtasks(js);
-  
-  if (entry->is_interval && entry->active) {
-    entry->target_time_ms = get_current_time_ms() + entry->interval_ms;
-  } else remove_timer(entry);
-  
-  current_time = get_current_time_ms();
-  ptr = &timer_state.timers;
-  goto scan;
-}
-
 int has_pending_timers(void) {
-  for (timer_entry_t *entry = timer_state.timers; entry != NULL; entry = entry->next) {
-    if (entry->active) return 1;
-  }
-  return 0;
+  return timer_state.active_timer_count > 0;
 }
 
 int has_pending_microtasks(void) {
   return timer_state.microtasks != NULL ? 1 : 0;
-}
-
-int64_t get_next_timer_timeout(void) {
-  uint64_t current_time = get_current_time_ms();
-  int64_t min_timeout = -1;
-  
-  for (timer_entry_t *entry = timer_state.timers; entry != NULL; entry = entry->next) {
-    if (!entry->active) continue;
-    
-    int64_t timeout = (int64_t)entry->target_time_ms - (int64_t)current_time;
-    if (timeout <= 0) return 0;
-    if (min_timeout == -1 || timeout < min_timeout) min_timeout = timeout;
-  }
-  
-  return min_timeout;
 }
 
 void init_timer_module() {
