+++ src/runtime.c
@@ -1,3 +1,5 @@
+#include "ant.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
